Restructure to address App -> Transforming classloader boundary.
In Phase 1, we load "PluginFile" resources to determine the service to
use for loading as well as sending up the Manifest to ModLauncher for
launch plugin interoperability (Mixin support) in Launcher.

In Phase 2, we load the resources again. We read in the metadata,
determine the candidates, and create actual plugins.

The Launch sequence is now determined. In Launcher, Phase 1 occurs in
the PluginDiscovererService that operates within App (see Phase 1
above). Phase 2 occurs when the specific sided "Launcher" is invoked
(i.e. ServerLauncher) which operates in Transforming (see Phase 2).

Remaining tasks:
- Determine final dependency versioning stategy
- Sort PluginCandidates per the strategy
- Finialize metadata and migrate to plugin-meta
- Create plugin object instances for the JVM language spi
- Separate out launcher code to SpongeCommon

Signed-off-by: Chris Sanders <zidane@spongepowered.org>